### 2018-09-03 0f44ae7
fixed typo in doc
