+++ man/draw_circle.Rd
@@ -4,7 +4,8 @@
 \alias{draw_circle}
 \title{Draw circle on image}
 \usage{
-draw_circle(im, x, y, radius, color = "white", opacity = 1, filled = TRUE)
+draw_circle(im, x, y, radius, color = "white", opacity = 1,
+  filled = TRUE)
 }
 \arguments{
 \item{im}{an image}
