+++ DESCRIPTION
@@ -13,7 +13,7 @@ Description: Fast image processing for images in up to 4 dimensions (two spatial
     dimensions, one time/depth dimension, one colour dimension). Provides most
     traditional image processing tools (filtering, morphology, transformations,
     etc.) as well as various functions for easily analysing image data using R. The
-    package wraps CImg, <http://cimg.eu>, a simple, modern C++ library for image
+    package wraps 'CImg', <http://cimg.eu>, a simple, modern C++ library for image
     processing.
 License: LGPL
 Imports:
@@ -25,7 +25,7 @@ BugReports: https://github.com/dahtah/imager/issues
 SystemRequirements: fftw3,libtiff,C++11
 LinkingTo: Rcpp
 LazyData: true
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
 Suggests:
     knitr,
     rmarkdown,ggplot2,dplyr,scales,
