### 2018-09-05 8eff3b8
added imdo, px.remove_outer
