+++ R/pixset.R
@@ -778,4 +778,18 @@ colorise <- function(im,px,col,alpha=1)
     imsplit(im,"c") %>% map2_il(col,mod) %>% imappend("c")
 }
 
-
+##' Remove all connected regions that touch image boundaries
+##'
+##' All pixels that belong to a connected region in contact with image boundaries are set to FALSE. 
+##' @param px a pixset
+##' @return a pixset 
+##' @author Simon Barthelmé
+##' @examples
+##' im <- draw_circle(imfill(100,100),c(0,50,100),c(50,50,50),radius=10,color=1)
+##' plot(im)
+##' as.pixset(im) %>% px.remove_outer %>% plot
+##' @export
+px.remove_outer <- function(px)
+{
+    pad(px,2,"xy",val=TRUE) %>% bucketfill(1,1,color=0) %>% crop.borders(1,1)  %>% as.pixset
+}
