+++ NAMESPACE
@@ -161,6 +161,7 @@ export(imappend)
 export(imchange)
 export(imcol)
 export(imdirac)
+export(imdo)
 export(imdraw)
 export(imeval)
 export(imfill)
@@ -238,6 +239,7 @@ export(px.flood)
 export(px.left)
 export(px.na)
 export(px.none)
+export(px.remove_outer)
 export(px.right)
 export(px.square)
 export(px.top)
