+++ appveyor.yml
@@ -1,57 +1,0 @@
-#uses code from https://raw.githubusercontent.com/clMathLibraries/clFFT/master/appveyor.yml
-
-# DO NOT CHANGE the "init" and "install" sections below
-platform:
-  - x64
-
-environment:
-  global:
-    LIB_FFTW: "%APPVEYOR_BUILD_FOLDER%\\bin\\fftw"
-
-# Download script file from GitHub
-init:
-  ps: |
-        $ErrorActionPreference = "Stop"
-        Invoke-WebRequest http://raw.github.com/krlmlr/r-appveyor/master/scripts/appveyor-tool.ps1 -OutFile "..\appveyor-tool.ps1"
-        Import-Module '..\appveyor-tool.ps1'
-
-install:
-  - ps: mkdir lib
-  - ps: pushd lib
-  - ps: echo Installing FFTW
-  - ps: mkdir $env:LIB_FFTW
-  - ps: pushd $env:LIB_FFTW
-  - ps: wget ftp://ftp.fftw.org/pub/fftw/fftw-3.3.4-dll64.zip -OutFile fftw-3.3.4-dll64.zip
-  - ps: 7z x fftw-3.3.4-dll64.zip
-  - ps: lib /machine:x64 /def:libfftw3-3.def
-  - ps: lib /machine:x64 /def:libfftw3f-3.def
-  - ps: lib /machine:x64 /def:libfftw3l-3.def
-  - ps: Bootstrap
-
-build_script:
-  - travis-tool.sh install_deps
-
-test_script:
-  - travis-tool.sh run_tests
-
-on_failure:
-  - travis-tool.sh dump_logs
-
-artifacts:
-  - path: '*.Rcheck\**\*.log'
-    name: Logs
-
-  - path: '*.Rcheck\**\*.out'
-    name: Logs
-
-  - path: '*.Rcheck\**\*.fail'
-    name: Logs
-
-  - path: '*.Rcheck\**\*.Rout'
-    name: Logs
-
-  - path: '\*_*.tar.gz'
-    name: Bits
-
-  - path: '\*_*.zip'
-    name: Bits
