+++ man/imdo.Rd
@@ -1,27 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/imeval.R
-\name{imdo}
-\alias{imdo}
-\title{imdo runs imeval, and reshapes the output as an image of the same dimensions as the input (useful for functions that return vectors).}
-\usage{
-imdo(obj, form)
-}
-\arguments{
-\item{form}{a single formula}
-}
-\description{
-imdo runs imeval, and reshapes the output as an image of the same dimensions as the input (useful for functions that return vectors).
-}
-\examples{
-#The rank function outputs a vector
-grayscale(boats) \%>\% rank \%>\% class
-#Auto-reshape into an image
-grayscale(boats)  \%>\% imdo(~ rank(.)) \%>\% plot
-#Note that the above performs histogram normalisation
-
-#Also works on lists
-imsplit(boats,"c") \%>\% imdo( ~ rank(.)) \%>\% imappend("c") \%>\% plot
-}
-\author{
-Simon Barthelme
-}
