### 2018-09-26 #919 @Kevin-Mattheus-Moerman
@jeroen the reviewer guidelines are here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

This is a pre-review issue and I aim to attract some more reviewers. Once I do I'll start the actual review issue and review process. Here is an [example of a review process](https://github.com/openjournals/joss-reviews/issues/890). As you can see the review process is guided by a set of tick boxes that will appear for you at the top of the review issue.
Let me know if you have more questions.
