### 2018-09-26 #919 @NikEfth
I do not have tons of experience with R but I could check the image processing aspects.
