### 2018-09-26 d003e0e
fixed help for split_connected
