### 2018-09-26 703f501
started removing some old cruft
