+++ R/utils.R
@@ -847,3 +847,15 @@ flatten.alpha <- function(im,bg="white")
         im
     }
 }
+
+
+#from ggplot2 by H. Wickham
+try_require <- function(package, fun) {
+  if (requireNamespace(package, quietly = TRUE)) {
+    library(package, character.only = TRUE)
+    return(invisible())
+  }
+
+  stop("Package `", package, "` required for `", fun , "`.\n",
+    "Please install and try again.", call. = FALSE)
+}
