+++ DESCRIPTION
@@ -17,7 +17,7 @@ Description: Fast image processing for images in up to 4 dimensions (two spatial
     processing.
 License: LGPL
 Imports:
-    Rcpp (>= 0.11.5),methods,stringr,png,jpeg,readbitmap,grDevices,purrr,Cairo,downloader,igraph,plyr
+    Rcpp (>= 0.11.5),methods,stringr,png,jpeg,readbitmap,grDevices,purrr,Cairo,downloader,plyr
 Depends:
     R (>= 2.10.0),magrittr
 URL: http://dahtah.github.io/imager, https://github.com/dahtah/imager
@@ -27,7 +27,6 @@ LinkingTo: Rcpp
 LazyData: true
 RoxygenNote: 6.1.0
 Suggests:
-    knitr,
-    rmarkdown,ggplot2,dplyr,scales,
-    testthat,OpenMPController,raster,spatstat
+    knitr,rmarkdown,ggplot2,dplyr,scales,
+    testthat,OpenMPController,raster,spatstat,igraph
 VignetteBuilder: knitr
