### 2018-10-01 #919 @dahtah
@Kevin-Mattheus-Moerman, I'm sorry, the people I know at the intersection of R and image processing are all imager users or contributors, I'm not sure they'd be appropriate as reviewers.
