### 2018-10-10 #919 @Kevin-Mattheus-Moerman
@dahtah contributors would have a potential conflict of interest. Users of the software however may not so they could be potential reviewers. I'll start the review process now but we can add new reviewers at a later stage as well.
