### 2018-10-10 #1012 @whedon
[REVIEW]: imager: an R package for image processing
**Submitting author:** @dahtah (<a href="http://orcid.org/0000-0003-0174-5058">Simon BARTHELME</a>)
**Repository:** <a href="https://github.com/dahtah/imager" target ="_blank">https://github.com/dahtah/imager</a>
**Version:** 0.42

- [x] **Version:** 0.42
