### 2018-10-23 #1012 @jeroen
@dahtah in the brief overview of imaging libraries available in R, could you compare `imager` to the `magick` package? This package interfaces to ImageMagick library and is the most comprehensive imaging toolkit in R currently available.
