### 2018-11-02 #1012 @NikEfth
@jeroen @dahtah I would need a bit of support to get this started. I am using arch linux and I cannot find the dependency ibfftw3-dev. Any help?
