### 2018-11-05 #1012 @dahtah
@NikEfth I don't know much about arch linux, but I guess [this](https://www.archlinux.org/packages/extra/x86_64/fftw/) should be the right package, assuming the include files are in there. Could you please confirm?
