### 2018-11-20 #1012 @Kevin-Mattheus-Moerman
@dahtah can you clarify the license for this work and where it can be found? Thanks.
