### 2018-11-21 #1012 @dahtah
@Kevin-Mattheus-Moerman The license can be found in the DESCRIPTION file of the R package, as per recommend practice. Do you mean clarify the license in the manuscript?
