### 2018-11-21 #1012 @Kevin-Mattheus-Moerman
Thanks for the clarification. You do not need to mention the license in the paper. But here are some steps I think you should take:

- [ ] The DESCRIPTION file simply reads: `License: LGPL`, however this license type has [three variations i.e. LGPL-2.0, LGPL-2.1 and LGPL-3.0](https://opensource.org/licenses/lgpl-license), which one does your project carry?

- [ ] Can you also add an actual LICENSE file to your project? If you choose LGPL-3.0 you'd paste [this text](https://opensource.org/licenses/LGPL-3.0) in there. This way it is clear what the license is. Also this way GitHub will render a little statement on the top right on your project front page clarifying the license, like this:
![screenshot from 2018-11-21 09-04-24](https://user-images.githubusercontent.com/8392709/48830493-a24ce780-ed6c-11e8-927c-5b33ce7a88d0.png)
