+++ DESCRIPTION
@@ -15,7 +15,7 @@ Description: Fast image processing for images in up to 4 dimensions (two spatial
     etc.) as well as various functions for easily analysing image data using R. The
     package wraps 'CImg', <http://cimg.eu>, a simple, modern C++ library for image
     processing.
-License: LGPL
+License: LGPL-3
 Imports:
     Rcpp (>= 0.11.5),methods,stringr,png,jpeg,readbitmap,grDevices,purrr,Cairo,downloader,igraph,plyr
 Depends:
