### 2018-11-26 #1012 @dahtah
@Kevin-Mattheus-Moerman Thank you for your suggestions. They are implemented in the current master.
