### 2018-11-28 #1012 @Kevin-Mattheus-Moerman
@arfon this repository is currently on the release 0.20 but the version label here and on CRAN is 0.41.1. Is this okay or should we suggest that the repo release carries the same version number as the CRAN number?
