### 2018-11-28 #1012 @Kevin-Mattheus-Moerman
@NikEfth have you resolved the linux arch issue you had? Are you still able to help with this review? Let me know if you cannot. Thanks.
