### 2018-11-28 #1012 @arfon
> @arfon this repository is currently on the release 0.20 but the version label here and on CRAN is 0.41.1. Is this okay or should we suggest that the repo release carries the same version number as the CRAN number?

I think we want to be clear about which version we've reviewed here. Is GitHub version `0.20` the same as CRAN version `0.41.1`?
