### 2018-11-28 #1012 @dahtah
@arfon It's my fault, I haven't consistently tagged releases on github to coincide with CRAN. If you'd like to know which version you have installed, run:
```packageVersion('imager')```
I'll try to pay more attention to tags in the future.
