### 2019-01-12 #1012 @Kevin-Mattheus-Moerman
@jeroen I see you have 3 unticked boxes remaining. The version number will be fixed once the submission is accepted. In terms of the license this has now been clarified to be an OSI approved license (LGPL-3.0). Are you able to tick the last boxes and perhaps formulate a recommendation (e.g. accept or not). Thanks!
