### 2019-01-25 e973d93
Fix fileext function pattern matching
Given the {3} quantifier in the expression, the current fileext function matches only three letters following the period, so it can extract the jpg file extension, but it can't extract the extension when jpeg is used:

> fileext <- function() { sub(".*\\.([^.]{3})$", "\\1", f) %>% tolower }
> fileext("myimage.jpeg")
[1] "myimage.jpeg"
> fileext("myimage.jpg")
[1] "jpg"

Therefore, when load.image for example checks to see if the output of fileext is png, jpg, jpeg, or bmp, it will raise an error for jpeg extensions, even though they are supported.

I modif
[…]
