### 2019-02-02 786ae5f
fixed minor warnings
