### 2019-03-01 #1012 @jeroen
OK accept.

PS: I don't understand why I have to check for ` Version: Does the release version given match the GitHub release (0.41.1)?` and then you say `
The version number will be fixed once the submission is accepted`. .
