### 2019-03-01 #1012 @Kevin-Mattheus-Moerman
Thanks @jeroen you are correct. In fact for future review issue we've removed the version check as we tend to update this anyway at the point of acceptance. Feel free to tick that last box. Thanks again for your review efforts!
