### 2019-04-10 #1012 @arfon
I *think* because it's missing this section: https://github.com/dahtah/imager/pull/91
