### 2019-04-18 7856ec4
Small typo on required header for Linux
The `libX11-dev` header is actually named `libx11-dev`. Just changing the `X` to lowercase and copying and pasting the command should work correctly.
