+++ NAMESPACE
@@ -1,274 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method("[",cimg)
-S3method("[",imlist)
-S3method("[",pixset)
-S3method("[<-",cimg)
-S3method("[<-",pixset)
-S3method(Ops,imager_array)
-S3method(as.array,cimg)
-S3method(as.cimg,"function")
-S3method(as.cimg,RasterLayer)
-S3method(as.cimg,array)
-S3method(as.cimg,cimg)
-S3method(as.cimg,data.frame)
-S3method(as.cimg,double)
-S3method(as.cimg,logical)
-S3method(as.cimg,matrix)
-S3method(as.cimg,numeric)
-S3method(as.cimg,pixset)
-S3method(as.cimg,raster)
-S3method(as.cimg,vector)
-S3method(as.data.frame,cimg)
-S3method(as.data.frame,imlist)
-S3method(as.data.frame,pixset)
-S3method(as.igraph,cimg)
-S3method(as.igraph,pixset)
-S3method(as.imlist,RasterStackBrick)
-S3method(as.imlist,cimg)
-S3method(as.imlist,imlist)
-S3method(as.imlist,list)
-S3method(as.list,imlist)
-S3method(as.logical,pixset)
-S3method(as.matrix,cimg)
-S3method(as.matrix,pixset)
-S3method(as.pixset,array)
-S3method(as.pixset,cimg)
-S3method(as.pixset,logical)
-S3method(as.pixset,pixset)
-S3method(as.raster,cimg)
-S3method(contours,cimg)
-S3method(contours,pixset)
-S3method(display,cimg)
-S3method(display,list)
-S3method(display,pixset)
-S3method(plot,cimg)
-S3method(plot,imlist)
-S3method(plot,pixset)
-S3method(print,cimg)
-S3method(print,imlist)
-S3method(print,pixset)
-export("%inr%")
-export("B<-")
-export("G<-")
-export("R<-")
-export("at<-")
-export("channel<-")
-export("color.at<-")
-export("frame<-")
-export(B)
-export(Cc)
-export(FFT)
-export(G)
-export(HSItoRGB)
-export(HSLtoRGB)
-export(HSVtoRGB)
-export(LabtoRGB)
-export(LabtoXYZ)
-export(LabtosRGB)
-export(R)
-export(RGBtoHSI)
-export(RGBtoHSL)
-export(RGBtoHSV)
-export(RGBtoLab)
-export(RGBtoXYZ)
-export(RGBtoYCbCr)
-export(RGBtoYUV)
-export(RGBtosRGB)
-export(XYZtoLab)
-export(XYZtoRGB)
-export(Xc)
-export(YCbCrtoRGB)
-export(YUVtoRGB)
-export(Yc)
-export(Zc)
-export(add)
-export(add.color)
-export(add.colour)
-export(as.cimg)
-export(as.imlist)
-export(as.pixset)
-export(at)
-export(autocrop)
-export(average)
-export(bbox)
-export(blur_anisotropic)
-export(boundary)
-export(boxblur)
-export(boxblur_xy)
-export(bucketfill)
-export(cannyEdges)
-export(capture.plot)
-export(center.stencil)
-export(channel)
-export(channels)
-export(ci)
-export(cimg)
-export(cimg.use.openmp)
-export(cimg2im)
-export(circles)
-export(clean)
-export(color.at)
-export(colorise)
-export(contours)
-export(convolve)
-export(coord.index)
-export(correlate)
-export(crop.bbox)
-export(crop.borders)
-export(depth)
-export(deriche)
-export(diffusion_tensors)
-export(dilate)
-export(dilate_rect)
-export(dilate_square)
-export(displacement)
[…]
