### 2019-05-15 #1012 @Kevin-Mattheus-Moerman
@dahtah can you add the following sections/headings to your paper:

- [ ] Acknowledgements

- [ ] References
