### 2019-05-30 #1012 @dahtah
Sorry for the delay, and thanks for reviewing this. Zenodo gave me the following DOI: 10.5281/zenodo.3235256. The archived version is 0.42
