### 2019-05-30 de4217e
Add plyr::rename as rename_plyr, convert llply to purrr::map and laply to purrr::map_lgl or purrr::map_dbl. Remove plyr deps from package.
