+++ R/utils.R
@@ -463,7 +463,7 @@ cut.kmeans <- function(x)
 ##' someFiles <- dir("*.png") #Find all PNGs in directory
 ##' iminfo(someFiles[1])
 ##' #Get info on all files, as a data frame
-##' info <- plyr::ldply(someFiles,function(v) iminfo(v) %>% as.data.frame) 
+##' info <- purrr::map_df(someFiles,function(v) iminfo(v) %>% as.data.frame)
 ##'}
 ##' @export
 iminfo <- function(fname)
@@ -846,4 +846,40 @@ flatten.alpha <- function(im,bg="white")
     {
         im
     }
+}
+
+
+#' Modify names by name, not position. (copied from plyr)
+#'
+#' @param x named object to modify
+#' @param replace named character vector, with new names as values, and
+#'   old names as names.
+#' @param warn_missing print a message if any of the old names are
+#'   not actually present in \code{x}.
+#' @param warn_duplicated print a message if any name appears more
+#'   than once in \code{x} after the operation.
+#' Note: x is not altered: To save the result, you need to copy the returned
+#'   data into a variable.
+#' @importFrom stats setNames
+#' @examples
+#' x <- c("a" = 1, "b" = 2, d = 3, 4)
+#' # Rename column d to "c", updating the variable "x" with the result
+#' x <- rename(x, replace = c("d" = "c"))
+#' x
+#' # Rename column "disp" to "displacement"
+#' rename(mtcars, c("disp" = "displacement"))
+rename_plyr <- function(x, replace, warn_missing = TRUE, warn_duplicated = TRUE ) {
+
+  # This line does the real work of `rename()`.
+  names(x) <- revalue(names(x), replace, warn_missing = warn_missing)
+
+  # Check if any names are duplicated.
+  duplicated_names <- names(x)[duplicated(names(x))]
+  if (warn_duplicated && (length(duplicated_names) > 0L)) {
+    duplicated_names_message <- paste0("`", duplicated_names, "`", collapse=", ")
+    warning("The imager::rename_plyr operation has created duplicates for the ",
+            "following name(s): (", duplicated_names_message, ")",
+            call. = FALSE)
+  }
+  x
 }
