### 2019-05-30 f5ed45a
Corresponding documentation updates
