### 2019-05-30 7ec8236
Corresponding documentation updates
