+++ NAMESPACE
@@ -292,13 +292,6 @@ importFrom(jpeg,readJPEG)
 importFrom(jpeg,writeJPEG)
 importFrom(magrittr,"%>%")
 importFrom(methods,is)
-importFrom(plyr,ddply)
-importFrom(plyr,dlply)
-importFrom(plyr,laply)
-importFrom(plyr,ldply)
-importFrom(plyr,llply)
-importFrom(plyr,mutate)
-importFrom(plyr,rename)
 importFrom(png,readPNG)
 importFrom(png,writePNG)
 importFrom(purrr,keep)
