### 2019-06-03 afdcf5f
plyr::mutate -> dplyr::mutate
