### 2019-06-03 6ec1864
Add pipe to package directly
