+++ NAMESPACE
@@ -48,6 +48,7 @@ S3method(plot,pixset)
 S3method(print,cimg)
 S3method(print,imlist)
 S3method(print,pixset)
+export("%>%")
 export("%inr%")
 export("B<-")
 export("G<-")
