### 2019-06-03 9b618ca
Documentation updates
