+++ man/imgradient.Rd
@@ -21,7 +21,7 @@ Light interface for get_gradient. Refer to get_gradient for details on the compu
 }
 \examples{
 grayscale(boats) \%>\% imgradient("x") \%>\% plot
-imgradient(boats,"xy") #Returns a list
+imgradient(boats,"xy") #Returns a list 
 }
 \author{
 Simon Barthelme
