+++ man/iiply.Rd
@@ -20,6 +20,6 @@ This is just imsplit followed by purrr::map followed by imappend
 }
 \examples{
 ##' #Normalise colour channels separately, recombine
-iiply(boats,"c",function(v) (v-mean(v))/sd(v)) \%>\% plot
+iiply(boats,"c",function(v) (v-mean(v))/sd(v)) \%>\% plot 
 
 }
