+++ man/imdraw.Rd
@@ -25,7 +25,7 @@ Draw image on another image
 \examples{
 im <- load.example("parrots")
 boats.small <- imresize(boats,.5)
-#I'm aware the result is somewhat ugly
+#I'm aware the result is somewhat ugly 
 imdraw(im,boats.small,x=400,y=10,opacity=.7) \%>\% plot
 }
 \seealso{
