### 2019-06-03 10e8c0c
Add mutate_plyr function, update documentation.
