+++ R/utils.R
@@ -5,7 +5,7 @@
 ##' Moisan (2011) defines an additive image decomposition
 ##' im = periodic + smooth
 ##' where the periodic part shouldn't be too far from the original image. The periodic part can be used in frequency-domain analyses, to reduce the artifacts induced by non-periodicity.
-##' 
+##'
 ##' @param im an image
 ##' @return an image
 ##' @examples
@@ -16,7 +16,7 @@
 ##' #The smooth error is the difference between
 ##' #the original image and its periodic part
 ##' (im-periodic.part(im)) %>% plot(main="Smooth part")
-##' 
+##'
 ##' @references  L. Moisan, Periodic plus Smooth Image Decomposition,J. Math. Imaging Vision, vol. 39:2, pp. 161-179, 2011
 ##' @author Simon Barthelme
 ##' @export
@@ -30,7 +30,7 @@ periodic.part <- function(im)
             {
                 periodic_part(im)
             }
-        
+
     }
 
 
@@ -38,13 +38,13 @@ periodic.part <- function(im)
 ##'
 ##' This function is equivalent to R's builtin fft, up to normalisation (R's version is unnormalised, this one is). It calls CImg's implementation.
 ##' Important note: FFT will compute a multidimensional Fast Fourier Transform, using as many dimensions as you have in the image, meaning that if you have a colour video, it will perform a 4D FFT. If you want to compute separate FFTs across channels, use imsplit.
-##' 
+##'
 ##' @param im.real The real part of the input (an image)
-##' @param im.imag The imaginary part (also an image. If missing, assume the signal is real). 
+##' @param im.imag The imaginary part (also an image. If missing, assume the signal is real).
 ##' @param inverse If true compute the inverse FFT (default: FALSE)
 ##' @return a list with components "real" (an image) and "imag" (an image), corresponding to the real and imaginary parts of the transform
 ##' @examples
-##' 
+##'
 ##' im <- as.cimg(function(x,y) sin(x/5)+cos(x/4)*sin(y/2),128,128)
 ##' ff <- FFT(im)
 ##' plot(ff$real,main="Real part of the transform")
@@ -69,17 +69,17 @@ FFT <- function(im.real,im.imag,inverse=FALSE)
     }
 
 
-##' Resize image uniformly 
+##' Resize image uniformly
 ##'
-##' Resize image by a single scale factor. For non-uniform scaling and a wider range of options, see resize. 
+##' Resize image by a single scale factor. For non-uniform scaling and a wider range of options, see resize.
 ##'
 ##' @name resize_uniform
 ##' @param im an image
 ##' @param scale a scale factor
-##' @param interpolation interpolation method to use (see doc for resize). Default 3, linear. Set to 5 for cubic, 6 for Lanczos (higher quality).  
+##' @param interpolation interpolation method to use (see doc for resize). Default 3, linear. Set to 5 for cubic, 6 for Lanczos (higher quality).
 ##' @return an image
 ##' @references
-##' For double-scale, triple-scale, etc. uses an anisotropic scaling algorithm described in: \url{http://scale2x.sourceforge.net/algorithm.html}. For half-scaling uses what the CImg doc describes as an "optimised filter", see resize_halfXY in CImg.h. 
+##' For double-scale, triple-scale, etc. uses an anisotropic scaling algorithm described in: \url{http://scale2x.sourceforge.net/algorithm.html}. For half-scaling uses what the CImg doc describes as an "optimised filter", see resize_halfXY in CImg.h.
 ##' @seealso resize
 ##' @examples
 ##' im <- load.example("parrots")
@@ -92,7 +92,7 @@ NULL
 ##' @export
 imresize <- function(im,scale=1,interpolation=3)
[…]
