### 2019-06-03 6d79316
Plyr function documentation
