### 2019-06-03 23ff234
Remove rename_plyr from documentation
