+++ man/rename_plyr.Rd
@@ -1,34 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{rename_plyr}
-\alias{rename_plyr}
-\title{Modify names by name, not position.}
-\usage{
-rename_plyr(x, replace, warn_missing = TRUE, warn_duplicated = TRUE)
-}
-\arguments{
-\item{x}{named object to modify}
-
-\item{replace}{named character vector, with new names as values, and
-old names as names.}
-
-\item{warn_missing}{print a message if any of the old names are
-not actually present in \code{x}.}
-
-\item{warn_duplicated}{print a message if any name appears more
-  than once in \code{x} after the operation.
-Note: x is not altered: To save the result, you need to copy the returned
-  data into a variable.}
-}
-\description{
-This function copied directly from plyr, and modified to use a different
-name to avoid namespace collisions with dplyr/tidyverse functions.
-}
-\examples{
-x <- c("a" = 1, "b" = 2, d = 3, 4)
-# Rename column d to "c", updating the variable "x" with the result
-x <- rename_plyr(x, replace = c("d" = "c"))
-x
-# Rename column "disp" to "displacement"
-rename_plyr(mtcars, c("disp" = "displacement"))
-}
