### 2019-06-03 4d2aea7
Don't link to outside functions (since the documentation doesn't exist)
