### 2019-06-03 1fef96a
Remove examples - not necessary for non-exported function
