### 2019-06-04 #1012 @Kevin-Mattheus-Moerman
@dahtah can you do the following:

- [ ] Can you change the Zenodo archive title from _"dahtah/imager: JOSS release"_ to match that of this submission _"imager: an R package for image processing"_

- [ ] Can you update the author list on Zenodo so it matches the paper?
