### 2019-06-21 e6c2a9e
added cimg2magick
