+++ NAMESPACE
@@ -107,6 +107,7 @@ export(ci)
 export(cimg)
 export(cimg.use.openmp)
 export(cimg2im)
+export(cimg2magick)
 export(circles)
 export(clean)
 export(color.at)
